dutzende Aenderungen. Siehe Changelog

+++ index.php
@@ -2,13 +2,13 @@
 $globvar = array();
 $globvar['title'] = 'IMGhost'; 
 $globvar['2ndtitle'] = 'Host you images';
-$globvar['imagesupport'] = 'jpg, png, gif';
+$globvar['imagesupport'] = 'jpg, png, gif'; // eigentlich schwachsinnig :p
 $globvar['maxsize'] = '1536'; // Angabe in Kilobyte
-$globvar['thumbwidth43'] = 180;  	/* neu resize */
-$globvar['thumbheight43'] = 135; 	/* für 16/9  */
-$globvar['thumbwidth169'] = 240; 	/* und 4/3 	*/
+$globvar['thumbwidth43'] = 180;  	// neu resize 
+$globvar['thumbheight43'] = 135; 	// für 16/9/
+$globvar['thumbwidth169'] = 240; 	// und 4/3
 $globvar['thumbheight169'] = 135; 
-$globvar['visiblecopyright'] = '<a href="http://www.somegas.de">Sascha Ohms</a>';
+$globvar['visiblecopyright'] = ' &copy;2009 <a href="http://www.somegas.de">Sascha Ohms</a>'; // Darf frei editiert werden
 $globvar['language'] = 'DE';
 $globvar['metadescription'] = 'Ein kostenloser Image-Hoster. Zeige deine Bilder, Freunden, in Foren oder irgendwo im Internet.';
 $globvar['use_randomname'] = true; // Zufallsname oder alten Dateinamen übernehmen
@@ -18,20 +18,24 @@ ob_start();
 include('tpl/header.tpl.php');
 
 if(isset($_GET['s'])) {
+	// Fragt ab, ob ueberhaupt der Bildname mitgeliefert wurde
 	if(empty($_GET['s'])) {
 		print '<p><img src="inc/img/zeichen.png" alt=""/> Kein Bild angegeben.</p>';
-	} elseif(!file_exists('./i/' .$_GET['s']) || is_valid_filename($_GET['s'])) {
+	// Prueft, ob das Bild existiert und ob der Bildname auch valide ist
+	} elseif(!file_exists('./i/' .$_GET['s']) && !is_valid_filename($_GET['s'])) {
 		echo '<p><img src="inc/img/zeichen.png" alt=""/> Datei existiert nicht oder ist unzul&auml;ssig.</p>';
 	} else {
 		$img = $_GET['s'];
 		$imgausgabe = './i/' .$img;
 		$clickfile = 'clicks/' .$img. '.cl';
 		
+		// Prueft, ob Click-Datei existiert. Wenn nicht wird diese erstellt
 		if(!file_exists($clickfile)) 
 			$count = 1;
 		else
 			$count = file_get_contents($clickfile)+1;
-			
+		
+		// Clickdatei auslesen und ausgeben
 		$fp = @fopen($clickfile, 'w');
 		@fwrite($fp, $count);
 		@fclose($fp);
@@ -49,28 +53,31 @@ if(isset($_GET['s'])) {
 		
 		$endung = substr($thename, -4);
 		
+		// Abfrage, ob Bildnamen nach dem Zufall gewaehlt werden sollen. Wenn nicht wird der eigentlich Bildname verwendet
 		if($globvar['use_randomname'] == true)
 			$name = substr(md5(uniqid(rand(), true)), 0, 12).$endung; 
 		else 
 			$name = $thename;
 		
+		// Bildgroesse in kb umrechnen
 		$size = $_FILES['nfile']['size']; 
 		$size = round($size / 1024, 2);
 		
+		// Abfrage ob Bild die maximale groesse ueberschreitet
 		if($size > $globvar['maxsize']) {
 			echo '<p><img src="./inc/img/zeichen.png" alt=""/> Das Bild ist gr&ouml;&szlig;er als ' .$globvar['maxsize']. ' kb</p>';
-		} elseif($type !== 'image/png' && $type !== 'image/x-png' && $type !== 'image/jpeg' && $type !== 'image/pjpeg' && $type !== 'image/gif') {
-			echo '<p><img src="./inc/img/zeichen.png" alt=""/> Dateityp wird nicht unterst&uuml;tzt</p>';
 		} else {	
 			$thumbdir = './i/t/';						
 			$thumb_width = $globvar['thumbwidth43'];
 			$thumb_height = $globvar['thumbheight43'];
 			
+			// Datei wird verschoben
 			move_uploaded_file($tempname, './i/' .$name);
-			$imginfo = getimagesize('./i/' .$name);
+			$imginfo = getimagesize('./i/' .$name); // Wichtige Abfrage fuer Bildtyp und Seitenverhaeltnisse
 			$height = $imginfo[1];
 			$width = $imginfo[0];
-
+			
+			// Entsprechend dem Bildnamen Funktionen ausaehlen
 			if($imginfo[2] == 2) {
 				$src = imagecreatefromjpeg('i/' .$name);
 				$typeausgabe = '.jpg';
@@ -82,13 +89,13 @@ if(isset($_GET['s'])) {
 				$typeausgabe = '.gif';
 			} else {
 				print '<p><img src="inc/img/zeichen.png" alt=""/> Dateityp wird nicht unterst&uuml;tzt.</p>';
-				@unlink($name);
+				@unlink($name); // Wichtig! Falls Dateityp nicht unterstuetzt wird, wird die Datei wieder geloescht
 				@unset($src);
 				include('tpl/footer.tpl.php');
 				exit;				
 			}
 			
-			/* 16/9 check und wenn kleiner als thumb resulotion kein resize */
+			// 16/9 check und wenn kleiner als thumb resulotion kein resize
 			if($width / $height >= 16 / 9) {
 				$thumb_width = $globvar['thumbwidth169'];
 				$thumb_height = $globvar['thumbheight169'];
@@ -102,9 +109,11 @@ if(isset($_GET['s'])) {
 				$thumb_height = $height;
 			}
 			
+			// Thumbnil generieren
 			$thumb = imagecreatetruecolor($thumb_width, $thumb_height);
 			imagecopyresampled($thumb, $src, 0, 0, 0, 0, $thumb_width, $thumb_height, $width, $height);
-				
+			
+			// Thumbnail mit richtigem Typ im Thumbnails-Ordner abspeichern
 			if($imginfo[2] == 2) 
 				imagejpeg($thumb, 'i/t/' .$name);
 			elseif($imginfo[2] == 3) 
@@ -112,14 +121,17 @@ if(isset($_GET['s'])) {
 			elseif($imginfo[2] == 1) 
 				imagegif($thumb, 'i/t/' .$name);							
 			
+			// Rechte auf Bild setzen. Muss nicht bei jedem Server gemacht werden, nur als Vorsichtsmassnahme
 			chmod('i/' .$name, 0644);
 			chmod('i/t/' .$name, 0644);
 			
+			// ServerURL auslesen
 			if(dirname($_SERVER['REQUEST_URI']) == "/")
 				$serverurl = $_SERVER['HTTP_HOST'].dirname($_SERVER['REQUEST_URI']);
 			else
 				$serverurl = $_SERVER['HTTP_HOST'].dirname($_SERVER['REQUEST_URI'])."/";
 			
+			// is.gd Link für Twitter generieren, wenn Twitter aktiviert ist
 			if($globvar['twitter'] == true) {
 				$isgdlink = isgd('http://' .$serverurl. 'i/' .$name);
 				if($isgdlink != "error")
@@ -127,7 +139,8 @@ if(isset($_GET['s'])) {
 				else
 					$twitterausgabe = 'http://twitter.com/home/?status=http://' .$serverurl. 'i/' .$name;
 			}
-				
+			
+			// Ausgabe Links definieren
 			$htmlcodeausgabe = '<img src="http://' .$serverurl. 'i/' .$name. '" alt="" />';
 			$bbcodeausgabe = '[img]http://' .$serverurl. 'i/' .$name. '[/img]';
 			$fullausgabe = 'http://' .$serverurl. 'i/' .$name;
@@ -142,11 +155,13 @@ if(isset($_GET['s'])) {
 }
 include('tpl/footer.tpl.php');
 
+// Funktion zum ueberpruefen, ob die Datei auch valide ist
 function is_valid_filename($filename, $extensions=array('jpg', 'jpeg', 'gif', 'png')) {
     $regex = '/^\w\.(' .implode('|', $extensions). ')$/';
     return preg_match($regex, $filename);
 }
 
+// Funktion fuer den API Call von is.gd
 function isgd($link) {
         $fp = fsockopen("www.is.gd", 80, $errno, $errstr, 30);
 	if (!$fp) {
@@ -159,7 +174,7 @@ function isgd($link) {
         	fwrite($fp, $out);
 		
 		while (!feof($fp)) {
- 	       		return substr(strstr(fread($fp, 300), 'http://'), 0, -5);
+ 	       		return substr(strstr(fread($fp, 300), 'http://'), 0, -5); // echt bescheiden geloest, aber es funktioniert :p
         	}
 		fclose($fp);
 	}
